+++ tests/data/model.sql
@@ -3,7 +3,7 @@ CREATE TABLE public.sample_model
     "id" bigserial PRIMARY KEY,
     "title" varchar(30) NOT NULL,
     "ids" bigint[],
-    "default_int" int DEFAULT 1,
+    "none_int" int,
     "created_datetime" timestamp,
     "ints" int[],
     "default_float" float DEFAULT 2.5,