+++ tests/test_table.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import List
+from typing import List, Optional
 import attr
 from attrs_to_sql.table import attrs_to_table
 
@@ -9,7 +9,7 @@ class SampleModel:
     id: int = attr.ib(metadata={"primary_key": True, "type": "bigint", "auto_inc": True})
     title: str = attr.ib(metadata={"not_null": True, "length": 30})
     ids: list = attr.ib(metadata={"type": "bigint[]"})
-    default_int: int = 1
+    none_int: Optional[int] = None
     created_datetime: datetime = attr.ib(factory=datetime.now)
     ints: List[int] = attr.ib(factory=list)
     default_float: float = 2.5
@@ -24,3 +24,4 @@ def test_attrs_to_table():
     actual_sql = attrs_to_table(SampleModel)
 
     assert actual_sql == expected_sql
+